add staticmethods to TripCreator class
and refactor existing creators
