+++ creators/accra/trips_creator_accra.py
@@ -3,21 +3,11 @@
 import sys
 import json
 import re
-import transitfeed
 from datetime import timedelta, datetime
 from creators.trips_creator import TripsCreator
 from core.osm_routes import Route, RouteMaster
 
-def _add_shape(schedule, route_id, osm_r):
-    shape_id = str(route_id)
-    try:
-        schedule.GetShape(shape_id)
-    except KeyError:
-        shape = transitfeed.Shape(shape_id)
-        for point in osm_r.shape:
-            shape.AddPoint(lat=float(point["lat"]), lon=float(point["lon"]))
-        schedule.AddShapeObject(shape)
-    return shape_id
+
 
 class TripsCreatorAccra(TripsCreator):
 
@@ -45,7 +35,7 @@ class TripsCreatorAccra(TripsCreator):
             route_index = 0
             for a_route_ref, a_route in line.routes.iteritems():
                 trip_gtfs = line_gtfs.AddTrip(schedule)
-                trip_gtfs.shape_id = _add_shape(schedule, a_route_ref, a_route)
+                trip_gtfs.shape_id = TripsCreator.add_shape(schedule, a_route_ref, a_route)
                 trip_gtfs.trip_headsign = a_route.to
                 trip_gtfs.direction_id = route_index % 2
                 route_index += 1
@@ -63,7 +53,7 @@ class TripsCreatorAccra(TripsCreator):
                     if index_stop == 0 :
                         trip_gtfs.AddStopTime(schedule.GetStop(str(stop_id)), stop_time=departure_time.strftime("%H:%M:%S"))
                     elif index_stop == len(a_route.stops) -1 :
-                        departure_time += timedelta(hours = 1) #TODO temps de trajet de bout en bout de chaque ligne à affiner
+                        departure_time += timedelta(hours = 1) #TODO
                         trip_gtfs.AddStopTime(schedule.GetStop(str(stop_id)), stop_time=departure_time.strftime("%H:%M:%S"))
                     else :
                         trip_gtfs.AddStopTime(schedule.GetStop(str(stop_id)))
@@ -74,14 +64,4 @@ class TripsCreatorAccra(TripsCreator):
                         stop_time.departure_secs = secs
                         trip_gtfs.ReplaceStopTimeObject(stop_time)
 
-                # interpolate times, because Navitia can not handle this itself
-                self.interpolate_stop_times(trip_gtfs)
-
-
-    @staticmethod
-    def interpolate_stop_times(trip):
-        for secs, stop_time, is_timepoint in trip.GetTimeInterpolatedStops():
-            if not is_timepoint:
-                stop_time.arrival_secs = secs
-                stop_time.departure_secs = secs
-                trip.ReplaceStopTimeObject(stop_time)
+                TripsCreator.interpolate_stop_times(trip_gtfs)
